requirements.txt
pyzmq
numpy
PyQt5
requests
click
tomli

